Fixed both issues:
1. 
add_course_sheet.dart:85 — _creditHours.toInt() returned int but creditHours expects double. Changed to pass _creditHours directly.
2. 
android/settings.gradle.kts — Updated AGP 8.2.1 → 8.6.0 and Kotlin 1.8.22 → 2.1.0 to match Flutter's minimum requirements.

+++ lib/presentation/widgets/add_course_sheet.dart
@@ -82,7 +82,7 @@ class _AddCourseSheetState extends ConsumerState<AddCourseSheet> {
       endTime: _timeToString(_endTime),
       colorValue: _colorValue,
       targetGrade: _targetGrade,
-      creditHours: _creditHours.toInt(),
+      creditHours: _creditHours,
       scheduleJson: _weekDays.map((d) => _days[d]).toList().toString(),
       weekDays: _weekDays.map((d) => _days[d]).toList(),
     );
